Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+2
> 
> Nice =) glad it was easy - do the functions work as expected ?
> Wrt. the icons - please poke Pedro.

Yes, both uno commands (of the new entries) work just like in desktop. The shape adjusts according to position of the cell in case of "anchor to cell". The shape adjusts to the size of the cell in case of "anchor to cell with scaling". 

But in desktop, there is an overlay image/icon of an anchor over the cell to which the shape is anchored to. I need to research a bit how to display that too. If I have to make up that on web-side (not as part of the tiles from core) then I'll poke Pedro for the icon when I have something working.

But should I wait for the above work to be complete before merging this patch ? Thanks for the review.

Patch-set: 1
Reviewer: Gerrit User 1000947 <1000947@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0