Update patch set 1
Patch Set 1:

A great question. Problem with the anchor overlay thing is that the anchor is I suspect shown per-view, and our tiles should always be for all users without such things. Also adding a client-side 'anchor rendering' layer is going to be rather expensive & slow, we have other things to fix now.

So lets leave that for now - I suspect the 'anchor to cell' does it to the nearest cell anyway ;-) so - should be easy enough for now.

Thanks !

Patch-set: 1