leaflet: double-tap to show keyboard on mobile
On mobile, we now do not show the keyboard by default
when the user taps the cell in calc. Instead, the user
should double-tap. This is especially useful when the
user selects the cell on first tap, then double-tap to edit.

On desktop and for other document the behavior is unchanged.

Change-Id: I4206291f60656391d061387f7bbad1b6e3caf80d