leaflet: double-tap to show keyboard on mobile
On mobile, we now do not show the keyboard by default
when the user taps the document. Instead, the user should
double-tap. This is similar to the behavior in Impress,
but especially useful for spreasheets, where the user
selects the cell on first tap, then double-tap to edit.

For Writer the behavior is slightly changed now, but
it should still be intuitive, since changing the cursor
position shows the keyboard. The only tricky part is that
double-tap also selects the word. We might want to special
case Writer if double-tapping isn't optimal and revert to
single-tap to edit.

On desktop the behavior should be unchanged.

Change-Id: I4206291f60656391d061387f7bbad1b6e3caf80d