WIP : Adapt align commands based on selection type
This patch works partly for Writer. It sends correct align command
based on map._clip._selectionType. But it still needs statechange
commands from core for the 'complex' selection type.

In Impress, this does not work at all as map._clip._selectionType
always says 'text' irrespective of what you are selecting. This is
pending investigation.

In Calc, aligning images/objects does not make sense w.r.t to the
sheet in general. The align buttons are disabled in core-desktop
when an image is selected.

Change-Id: I8223cd03f864fa92523224bf88ccb992edfc2686