Adapt align commands based on selection type
Use map._clip._selectionType to decide what align command to send
when one clicks on one of the toolbar align buttons. selectionType
is either 'text' or 'complex'; 'complex' by definition is set for
non-text selections like images and charts.

What is to be done next ?

1. Need to setup statechange: messages from core.git for
.uno:ObjectAlignLeft, .uno:AlignCenter, .uno:ObjectAlignRight
from sw/ sd/ in Get*State() methods of the relevant shells. After
this we need to make online toolbar align buttons to react to the
correct statechange messages based on selectionType.

2. _selectionType is not always correct. In Impress it
is always 'text' regardless of selection type. In Writer, for
shape selection, it is set to 'text'. SelectionType comes from
XTransferable2::isComplex(). Its implementations in sw/ and sd/
in core.git need fixing.

PS: In Calc, aligning images/objects does not make sense w.r.t to the
sheet in general. The align buttons are disabled in core-desktop
when an image is selected.

Change-Id: I8223cd03f864fa92523224bf88ccb992edfc2686