tdf#128468: Make dragging the tab stops send a UNO command
Still work in progress. The new UNO command is "MoveTabstop" with two
arguments: Tabstop (the index of the tabsetop) and Amount (in twips,
left (negative) or right).

The intent is that moving a tabstop in core will cause a callback that
will redraw the ruler, like moving a paragraph margin does.

Change-Id: Ic6b8b11811d2a7f6510c505fff421e1796300c70