mobile: Hide new style button on mobile wizard.
It opens a tunneled dialog, what we try to avoid on mobile.
Also after the recent focus related changes this dialog
can't get the focus, so can't add the name of the new style.
Even if I workround this focus issue and I manage to add a new
style, then that style won't be part of the style list displayed
on the mobile wizard.
All in all adding new style does not work now, so better to hide
this item.

Change-Id: I35c3686129621656f3f63914c9590248f320c15d