Fix app crash when re-opening a document quickly
Normally opening a document after closing one waits for
the activity to be destroyed However, when this is done quick enough,
it tries to load the document before the old one is destoryed and
it can be seen on the console by putting log under onDestroy method
after sending "BYE" message to native part. This crashes the app.
This patch is meant to prevent that from happening
altough it might not be ideal, but it works on the device i use.

Change-Id: I0a0cfdb07e6a2e9188cae4760e4658a03bc75358