zoom-adjust header-height only when we have edit permissions
The patch introduced adjusting column header height with zoom level

commit ab64d2e0c315d26c358ff99548bb7353708eab6b
Author: Dennis Francis <dennis.francis@collabora.com>
Date:   Wed Jan 29 13:38:16 2020 +0530

    Adjust column-header height with zoom-level too

That patch mess up the positions of headers in the below scenario:
1. First zoom-in or out in mobile mode before enabling edit permissions
2. Then enable edit permissions, and now the header positions are all
messed up.

This happens because the formulabar ui is show only while in edit-mode
and the resize() method is not very robust(it adjusts 'top' style attribute
with height deltas) and does not consider the introduction of formula-bar
div into account when one changes the permission of the sheet from
'readonly' to 'edit' in mobile mode.

So lets not do resize() of the column-header container before entering
edit mode and avoid complications of adjusting for the introduction of
the formula-bar.

Change-Id: Ib03f89ad7a34cbd2a3cbfb65af2a9a016b47c125