Update patch set 4
Patch Set 4:

I'm really not a fan of passing random parameters to map.focus().

I want a method we can call from anywheere:

map.focus()

And this un-screws-up the browser focus comprehensively. Perhaps it calls each docLayer to see if it thinks we should have it or something - but there is no: "did we miss a call-site", "oh, some state change happened that we missed" - no it is 10000% accurate - every time you call it it does the same thing for the given document state.

If we need to change the state- fine, but we have a nice descriptively named variable that has that inside it that documents why, and can be searched for to see when that state changes.

Or did I mis-read the patch ?

Patch-set: 4