Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> I'm really not a fan of passing random parameters to map.focus().

That is temporary. Because focus() is called from a large number of places, some of which are unnecessary/wrong, I didn't want to break the world and start from scratch (also, I couldn't test all possible states/features). Instead I override with that arg, leaving the default as-is. Once we have something sensible, we can remove the argument with explicit calls.

> And this un-screws-up the browser focus comprehensively. Perhaps it calls each docLayer to see if it thinks we should have it or something - but there is no: "did we miss a call-site", "oh, some state change happened that we missed" - no it is 10000% accurate - every time you call it it does the same thing for the given document state.

Currently there is no centralized state, for better or worse. Some states are tracked (e.g. graphic selection) while others are handled as they change (i.e. most events coming from Core). But the worst ones are those that are tracked but unreliable (e.g. cursor visibility, which can change rapidly/randomly and is therefore not reliable to show/hide the keyboard).

So the current approach is to minimize the places we do map.focus() so that we go towards a more coherent state management.

My ultimate goal would be to have events fired for the different state-change notifications and have a centralized handler or handlers do the keyboard show/hide logic, as well as cursor show/hide etc. This should make the code more readable and maintainable, but to get there we need calculated changes that don't break the world (harder done than said).

Patch-set: 4
Reviewer: Gerrit User 1000871 <1000871@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0