Update patch set 4
Patch Set 4:

Hi Ash,

Lets get this in, but from a design perspective:

> Currently there is no centralized state, for better or worse.

That would be for worse of course. I want stored state in the JS that shadows (and is a read-only copy of) the core office state.

> Some states are tracked (e.g. graphic selection) 

Graphic selection works reliably.

> while others are handled as they change (i.e. most
> events coming from Core).

And many of these do not.

> But the worst ones are those that are tracked but unreliable
> (e.g. cursor visibility, which can change rapidly/randomly
> and is therefore not reliable to show/hide the keyboard).

We should track and store and maintain the state we get from the core here.

> So the current approach is to minimize the places we do map.focus()
> so that we go towards a more coherent state management.

Hmm. I am convinced that as long as we believe anything the browser says: "I lost focus" or whatever and act on these events we will loose, loose hard, and loose repeatedly. Instead we should repeatedly, and regularly kick the browser into doing what we know is the right thing.

> My ultimate goal would be to have events fired for the different
> state-change notifications

I don't care about the mechanics. I only care that we do no state 'edge' detection whereby we synchronize state from the core into the browser: since the browser will -inevitably- discard / throw away / trample on that state. Instead we we should store any state we get from the core - via events - and at any time know what we want to do wrt. focus / cursor handling etc. in the browser. Possibly this means also tracking the few native cursor things we need like comments / search bars etc.

> and have a centralized handler or handlers do the keyboard show/hide
> logic, as well as cursor show/hide etc. This should make the code
> more readable and maintainable, but to get there we need calculated
> changes that don't break the world (harder done than said).

Again. Can you confirm that we are moving to a world where we listen to the core events, we store these events in our own JS structures, and we then call a "makeTheBrowserDoIt()" method - that kicks the browser repeatedly until it does the right thing using only state that we trust (ie. from the core), and not "is something focused" from the browser itself (?) =)

Thanks.

Patch-set: 4
Reviewer: Gerrit User 1000037 <1000037@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Style=0