Update patch set 1
Patch Set 1:

> I wish we can use distinguishing names for these two cases. Is it obvious to everyone but me what each of these two flags represent? Is one more generic than the other? Then we should only use the more generic here, I guess. (Sorry if I missed some documentation somewhere.)

In general, whenever the code use the term "mobile" by itself without further qualification (as in "window.mode.isMobile()"), it means "mobile phone", because mobile phones is what people other than me are targeting and testing on. window.ThisIsAMobileApp again means that it is the iOS or Android app that the code is running on, and that flag is essentially a constant, set when loleaflet.html is produced at build time for the app from loleaflet.html.m4. It is true both on phones and tablets. Then there is also window.ThisIsTheiOSApp and window.ThisIsTheAndroidApp which are similar build-time constants with obvious meaning.

Should this be simplified or clarified? Perhaps, but I fear that if for instance that isMobile() is renamed to isMobilePhone(), it will firstly cause trouble when cherry-picking from other branches, and secondly it will take weeks before everybody notices that the API has been renamed. Another approach might be to make isMobile() return true also on tablets, not just phones. That again brings the risk that there might well be cases where isMobile() actually *is* used for cases that should be used on a phone, but not on a tablet.

Patch-set: 1
Reviewer: Gerrit User 1000062 <1000062@6622c302-437a-49ad-a91d-55ba7dcf1615>