leaflet: fix searching in document functionality
This makes searching work again, but typing 'r'
still triggers F3, which simply matches the next
hit (on desktop, on mobile it works fine).

So (on desktop) typing 'your' will match the first 'you'
and the 'r' will move the search results to the
second 'you' (assuming there are at least two 'you'
entries in the document and one 'your').

Change-Id: Iab425afdfb3848ecff97f5599ceb7cae160b097d