Revert "leaflet: Impress toolbar fixes for sidebar state"
This reverts commit bcd7af96579181969f1d6c1c977731e030fe3c6e.

The commit added the extra generic Sidebar toolbar button in Impress,
which had a conflicting state with the Slide Layout one.

Additionally, after hiding the sidebar with the specific button, and
trying to bring it back using the generic one, it can't be brought
back again in the same session with any of the sidebar toolbar buttons.

Change-Id: If5def8b4169c9721c863f935ff59127af522f376