Update patch set 1
Patch Set 1:

About the test failure. A lot of tests use the context menu for copying the selected text to the clipboard. The copying code uses '.ui-header.level-0.mobile-wizard.ui-widget .menu-entry-with-icon .context-menu-link' as a css selector, which is not working anymore, I guess. Maybe '.context-menu-link' can be the new selector.
However, as I see the table related tests are not failing, which also use the copy item from the context menu. Can it happen, that this change does not affect the table context menu?

Patch-set: 1
CC: Gerrit User 1000169 <1000169@6622c302-437a-49ad-a91d-55ba7dcf1615>