Instead of zooming each time we resize, instead keep zoom and pan.
This helps us smoothly follow the cursor, and to adapt to less
space available when eg. the wizard or the keyboard pop up.

Ensure we do the right zoom-to-fit on the first call, interestingly
if we do not do that - we get just a single tile at the top left (que?).

Change-Id: Ib26f9b474caa631028e18e790dd50c058cbaef3b