Lets have Hammer pan listeners on scroll-container...
and delegate the pan events to L.Map.TouchGesture's pan handlers.

This solves the problem that if one starts to pan the sheet from
near the right edge (where the vertical scroll div exists), no
panning takes place. This is because the vertical scroll div
is managed by custom jquery plugin and has a high z-index and
mapPane is not an ancestor of it, so mapPane will not get the
touch events even with event bubble-up. But its ancestor div
'scroll-container' still gets touch/move events and has the same
size and location as that of mapPane. So we just need to register
Hammer for pan events for scroll-container and delegate these events
to the original pan handlers of mapPane without any change to the
event coordinates.

The scroll widget and the drag-rail still function as usual as they
are succesors of scroll-container and they get all the events they
need to handle.

Change-Id: Ie4ffe07a0889c5710b2c6d09e4eb90f0671b5ad0