Update patch set 3
Patch Set 3:

Patch set 3 fixes a compilation problem for iOS.

I dropped the #error. It asked "Why are there IOS conditionals in Kit.cpp::lokit_main - and we don't use that ?" The answer is that we do. lokit_main is called from the thread started in getNewChild_Blocks(). (This is both on Android and on iOS.)

Patch-set: 3
Reviewer: Gerrit User 1000062 <1000062@6622c302-437a-49ad-a91d-55ba7dcf1615>