Add Label for the line-style listbox in jsdialog
The Line-style listbox selector(shape properties) in mobile online
is missing a label, so to make it clear lets add a label for it.
Adding an invisible label for line-style listbox in core.git does
not work as invisible widgets are not dumped by core to generate
jsdialog messages. So the next option is to splice in a label
entry to the parsed jsdialog message at the appropriate place.

This also makes it easier for adding any missing labels, by just
needing to add the id of the control and the label text.

Change-Id: I45913ec25278e8566092a738c08cfdd8bc46e39f