cypress: mobile: Fix-up hidden style items test.
Without waiting for the mobile wizard to be opened
the not exist check will allways pass no matter
whether the items are actually hidden or not.
First we need an indicator that the mobile wizard
is opened and then we can check whether the removed
styles items are not there.

Change-Id: I00c1aaf6e11351cf8714be3ed6d5b9d2ff000324