Update patch set 3
Patch Set 3:

It's actually an loolwsd option, something like:

 --o:trace[@enable]=true --o:trace.path=/tmp/debug/looltrace --o:trace.path[@compress]=false

gives me a trace file. The tricky part is to find a good balance: the fuzzed area should be small enough that the fuzzer can explore interesting paths quickly, e.g. Chromium says:

https://chromium.googlesource.com/chromium/src/testing/libfuzzer/+/refs/heads/master/efficient_fuzzing.md#execution-speed

> You should aim for at least 1,000 exec/s from your fuzz target locally before submitting it to the Chromium repository.

But at the same time if we fuzz only a oneliner function, that's too simple and not interesting, either. A challenging trade-off. :-)

So I guess this can be approached from two ends:

1) Either continue from the loolwsd_fuzzer direction, but simplify it further: it is already unprivileged and it already does not chrooting, but it still calls into LOK and it still forks, still does notwork communication, etc. A lot of work to cut this down to the 1ms/run speed.

2) Start from the other end. I guess our weakest point is the WSD-level (not Kit-level) websocket incoming traffic parser, so we could build a unit-test like executable that more or less directly feeds the input from the fuzzer into ClientSession::handleMessage().

(The fuzzer always gives us a byte array.)

My plan is to give 2) a try and see how it works. If succeeded, we can add more fuzzers, e.g. one that targets ChildSession, etc.

Oh and I think it may be easier to go with libfuzzer from LLVM (or at least start with that one), the core.git fuzzers use that, and not AFL. As far as I understand, they are similar, just "we" have a bit more experience with libfuzzer.

Patch-set: 3