Update patch set 1
Patch Set 1:

Would be great not to break the iPad while playing with Phablets (which is not really a critical task).

Clearly we need a good way to differentiate these - what is that ? can we use the DPIscale to work out what the real physical size of the screen is or ? ... but seems a really poor time to disturb this code.

Patch-set: 1
CC: Gerrit User 1000037 <1000037@6622c302-437a-49ad-a91d-55ba7dcf1615>