loleaflet: makefile: use m4 to bundle all js files
It is a flexible bundling of all javascript libraries and
each platform (ANDROID, IOS, Desktop) can remove or add
from node_modules or imported according to its requirements.

Maybe in a near future some external libraries could be loaded
on demand or lazy loading that are not necessary when the page
is loading.

Change-Id: Ief5d4973ab29a4fb3c4848c3e1653a24fff0da69