leaflet: tolerate larger multi-tap threshold
When double-tapping (or tripple) the user is almost
always going to place their finger at different
coordinates for each of the taps. There is a default
threshold of 10 units between the taps in Hammer
to tolerate for this error when detecting multiple
taps, rather than individual taps.

This threshold of 10 is really small and unrealistic
for our needs, as in most cases double-tapping (which
we use very often to select words and begin editing
in Impress) is not detected (single-taps are issued
instead). This makes interaction/usability very hard.

Here we increase the threshold to a comfortable 100.
If this causes any other usability issues, we can
lower it acccordingly, but so far it only improved
the experience for me.

Change-Id: I3fee43c3efaa4507cae09b7c564cb4c0c1b5384e