Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> it still fails somehow in the insert table..

You can ignore the insert table test case now. I see a different issue with that caused by another commit. It's disabled for Jenkins, so you can ignore it for now. I'll update it.
The issue why Jenkins fails with this patch is the Edit menu item. You changed the css selector for this menu item in helper.js, but the Edit menu item does not have an icon. I guess it's your other patch which adds icons for the hamburger menu.

Patch-set: 1
CC: Gerrit User 1000169 <1000169@6622c302-437a-49ad-a91d-55ba7dcf1615>