Don't bind the slideshow downloading to the richdocuments...
... decide based on the DownloadAsPostMessage wopi value instead.

The idea is that the Nextcloud app actually sets the DownloadAsPostMessage
wopi value to true, which means that all the downloading (including
slideshow) is passed as Download_As postMessage anyway.

So deciding here based on the existence of the Nextcloud app is
not necessary, because there might be others who set the
DownloadAsPostMessage and rely on the Download_As postMessage,
but would get a non-functional iframe over the top of the editing
window.

Change-Id: I4a5c1d9ef6276908a1b3910a4269091955cdc490