Update patch set 3
Patch Set 3: Code-Review-1

Mert: Thank you for the update!  From the code reading perspective, I wonder, the setPermission('edit') looks like enabling the editing mode, shouldn't it be setPermission('readonly') or ('view') or something (in both cases it the patch)?  Or am I just confused?

Also when I was testing this yesterday, we should definitely include also changing the shape of the 'back' in the top left to a checkmark, without that it feels a bit weird from the user point of view - "why do I need to tap the back twice to return?" :-)

I'm -1'ing for the moment, but love the direction of this & looking forward to the final patch - thank you for your work on this!

Patch-set: 3
Reviewer: Gerrit User 1000068 <1000068@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=-1