leaflet: don't lose focus when the dialog cursor is hiden
The cursor visibility can change rapidly, and the formula-bar
seems to do this when typing into it. The result is that
because of losing the input focus we lose some of the input.

Instead, we only use the dialog cursor update to show the keyboard
rather than to hide it. For hiding, other messages will do it.

Change-Id: I0e8674170aa9d05b6ddc15de8d35792377b6725b