Fix js crash when jsdialog invoked on piecharts
For some reason, the parsed jsdialog message for piecharts
in calc has element(s) with children that evaluates to undefined.
Although we need to fix why there is such an child for any ui object,
this is a quick fix that skips such entries, so that the chart dialog
is shown without a crash in the console. Else the dialog wont appear.

Change-Id: Iff11051bf775ece63ac9f0ca180b3c76d7d14e84