Update patch set 1
Patch Set 1: Code-Review+2

Looks much safer =) and prettier.

I'd really prefer to avoid heap allocation where it is easy; so this

getToken(tokens, "foo") - is good - but if we can avoid copying a std::string out to compare it to "foo" and then free it that's really good.

Also if we can have getInteger(tokens, "foo") or whatever with the conversion done withotu allocation that's generally good.

Beyond that - I guess that having standard JSON for all our messages would be nice ;-) but lets get this in 1st and iterate.

Thanks !

Patch-set: 1
Reviewer: Gerrit User 1000037 <1000037@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=+2