leaflet: resize the document after show/hide toolbar
The toolbar changes the document size, so we should resize
at the time of showing/hiding the toolbar.
Failing to do this means the document is resized at
the point of editing it, which causes a visual
shift as the contents recenters.

This avoids it by making sure the document dimensions
are up-to-date at the time the toolbar shown/hidden.

Change-Id: I728d13e1acac4b2dd118332a112d4b89b4d798b1