cypress: new impress editing tests
Single- and double-click starts editing
text-boxes in Impress. Single-clicking works
only on text, not in the white-space.

Merges an existing test with a new one to
minimize redundancy where there is overlap.

Improved helpers to minimize duplication
and update a single place when there are changes
to the common operations/checks.

Change-Id: I14ee65b18e310f4ff6b93ce81580b3d6aba0fa03