Remember the horizontal scroll position of calc tab-bar
Whenever the tab-bar contents are updated, restore the last
known scroll position.

This fixes a very annoying problem while selecting a sheet in
a document with many many sheets. Everytime the user selects a sheet
(closer to the last sheet), the tab-bar's scroll position is reset
to sheet#1, so if the user wants to select another sheet near the
last sheet, they have to scroll again a lot to see it.

Change-Id: I7e79b6e1a7b19223e1dbd9b3aab6c6f2b4d422b2