ChildSession::renderShapeSelection: vector -> unique_ptr
Open Writer, insert a multi-MP JPEG, select it.

Before:

ChildSession::renderShapeSelection: finished in 81.33 ms

After:

ChildSession::renderShapeSelection: finished in 74.67 ms (91.81% of baseline)

This is with an --enable-symbols core with a -O2 online, with libstdc++.

The cost on the Android profile with its libc++ looked even more,
spending time in the std::vector dtor.

Change-Id: I50af2e13fd24569dc32304420b8f3e70d15803eb