Update patch set 1
Patch Set 1: Code-Review-2

Glad it speeds up the unit tests ! =) that's cool in itself.

Problem is - this breaks the 'flushTimeoutMS' concept by creating a new flush-timeout which is 4x some small number.

It would be far better IMHO to have a method to check all of the Poll's sockets: "hasPending()" or something that checks that _outBuffer and _inBuffer are empty and also checks that if the ProtocolHandlerInterface is set then _proto->hasQueuedMessages() == false.

Can you do that instead ? =) hopefully that will also finish ~instantly and not give a need for any timeout at all in most cases ...

Patch-set: 1
Reviewer: Gerrit User 1000037 <1000037@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=-2