Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> (3 comments)
> 
> Great stuff, thank you for this!  Conceptually, I'd slightly prefer filtering out the removed Uris directly in addDocumentToRecents() though - we already shuffle the list there, see the "remove string if present, so that it doesn't appear multiple times" - so if we could check for validity some simple way already there, that would be my preferred.
> 
> But OK to do it the way in the patch too; just a bit of nitpicking below :-)

addDocumentToRecents() is called when the document is tried to be opened which is the problem. If the file uri is not valid, the app crashes :) also the file name looks like home:<document.odt> which does not look good.

Patch-set: 2
Reviewer: Gerrit User 1001306 <1001306@6622c302-437a-49ad-a91d-55ba7dcf1615>