Update patch set 8
Patch Set 8:

That's true that Jenkins uses core/master.

Anyway, it does not worth spending too much time on this, since the focus is still unreliable sometimes. For example, in Writer I see that double-clicking on a shape grabs the focus to the shape, but the blinking cursor is not visible (maybe that happens on Jenkins in Impress too). Actually first it is displayed for a short time and then it is hidden again. (tested on staging). I expect that later if things become more stable around focus/blinking cursor, we'll have less problem with the stability of cypress tests.

Patch-set: 8